+++ core/src/main/scala/org/apache/spark/sql/crossdata/catalog/XDCatalog.scala
@@ -15,6 +15,7 @@
  */
 package org.apache.spark.sql.crossdata.catalog
 
+
 import org.apache.spark.sql.catalyst.TableIdentifier
 import org.apache.spark.sql.catalyst.analysis.Catalog
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
@@ -27,17 +28,29 @@ import org.json4s.jackson.Serialization._
 
 object XDCatalog extends CrossdataSerializer {
 
-
 implicit def asXDCatalog (catalog: Catalog): XDCatalog = catalog.asInstanceOf[XDCatalog]
 
   type ViewIdentifier = TableIdentifier
 
+  case class IndexIdentifier(indexType: String, indexName: String) {
+    def quotedString: String = s"`$indexName`.`$indexType`"
+    def unquotedString: String = s"$indexName.$indexType"
+    override def toString: String = quotedString
+    def asTableIdentifier: TableIdentifier = TableIdentifier(indexType,Option(indexName))
+  }
 
   case class CrossdataTable(tableName: String, dbName: Option[String], schema: Option[StructType],
                             datasource: String, partitionColumn: Array[String] = Array.empty,
                             opts: Map[String, String] = Map.empty, crossdataVersion: String = crossdata.CrossdataVersion)
 
+
+  case class CrossdataIndex(tableIdentifier: TableIdentifier, indexIdentifier: IndexIdentifier,
+                            indexedCols: Seq[String], pk: String, datasource: String,
+                            opts: Map[String, String] = Map.empty, crossdataVersion: String = crossdata.CrossdataVersion)
+
+
   case class CrossdataApp(jar: String, appAlias: String, appClass: String)
+
 
   def serializeSchema(schema: StructType): String = write(schema)
 
@@ -50,6 +63,11 @@ implicit def asXDCatalog (catalog: Catalog): XDCatalog = catalog.asInstanceOf[XD
   def serializeOptions(options: Map[String, String]): String =  write(options)
 
   def deserializeOptions(optsJSON: String): Map[String, String] = read[Map[String, String]](optsJSON)
+
+  def serializeSeq(seq: Seq[String]): String = write(seq)
+
+  def deserializeSeq(seqJSON: String): Seq[String] = read[Seq[String]](seqJSON)
+
 
 }
 
@@ -67,8 +85,11 @@ with StreamingCatalogAPI {
   def unregisterView(viewIdentifier: ViewIdentifier): Unit
 
   /**
-    * Check the connection to the set Catalog
-    */
+   * Check the connection to the set Catalog
+   */
   def checkConnectivity: Boolean
 
 }
+
+
+