+++ core/src/main/scala/org/apache/spark/sql/crossdata/daos/impl/IndexTypesafeDAO.scala
@@ -13,20 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.spark.sql.crossdata.daos
+package org.apache.spark.sql.crossdata.daos.impl
 
-import java.util.UUID
+import com.typesafe.config.Config
+import org.apache.spark.sql.crossdata.daos.IndexDAO
 
-object DAOConstants {
+class IndexTypesafeDAO(configuration: Config) extends IndexDAO {
 
-  val BaseZKPath = "stratio/crossdata"
-  val TablesPath = s"$BaseZKPath/tables"
-  val ViewsPath = s"$BaseZKPath/views"
-  val AppsPath = s"$BaseZKPath/apps"
-  val IndexesPath = s"$BaseZKPath/indexes"
-  val EphemeralTablesPath = s"$BaseZKPath/ephemeraltables"
-  val EphemeralTableStatusPath = s"$BaseZKPath/ephemeraltablestatus"
-  val EphemeralQueriesPath = s"$BaseZKPath/ephemeralqueries"
+  override val config = new TypesafeConfig(Option(configuration))
 
-  def createId: String = UUID.randomUUID.toString
 }