+++ elasticsearch/src/main/scala/com/stratio/crossdata/connector/elasticsearch/DefaultSource.scala
@@ -185,16 +185,21 @@ class DefaultSource extends RelationProvider with SchemaRelationProvider
   override def dropExternalTable(context: SQLContext,
                                  options: Map[String, String]): Try[Unit] = {
 
-    val (index, typeName) = ElasticSearchConnectionUtils.extractIndexAndType(options).get
-    val indexType = IndexAndType(index, typeName)
+    if(ElasticSearchConnectionUtils.numberOfTypes(options) == 1) {
+      val (index, typeName) = ElasticSearchConnectionUtils.extractIndexAndType(options).get
+      val indexType = IndexAndType(index, typeName)
 
-    Try {
-      ElasticSearchConnectionUtils.withClientDo(options){ client =>
-        client.execute {
-          deleteIndex(index)
-        }.await
+      Try {
+        ElasticSearchConnectionUtils.withClientDo(options){ client =>
+          client.execute {
+            deleteIndex(index)
+          }.await
+        }
       }
+    } else {
+      sys.error("Cannot remove table from ElasticSearch if more than one table is persisted in the same index. Please remove it natively")
     }
+
   }
 
 }